requirements.txt
Django==1.5.2
argparse==1.2.1
distribute==0.6.24
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
django-vanilla-views==1.0.2
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
South==0.8.2